+++ src/eth/executor/types/mod.rs
@@ -1,7 +1,6 @@
 mod account_changes;
 mod error;
 mod execution_changes;
-mod execution_metrics;
 mod execution_result;
 mod task;
 mod transaction_execution;
@@ -13,8 +12,6 @@ pub use execution_changes::AccountOriginalsReader;
 pub use execution_changes::Complete;
 pub use execution_changes::ExecutionChanges;
 pub use execution_changes::Incomplete;
-pub use execution_metrics::EvmExecutionMetrics;
-pub use execution_metrics::SlotAccessMetrics;
 pub use execution_result::ExecutionResult;
 pub use execution_result::RevertReason;
 pub use task::EvmRoute;
